// src/types/checkoutPage.enum.ts
export enum CheckoutPageTexts {
  PrimaryHeader = "Swag Labs",
  SecondaryHeader = "Checkout: Your Information",
  SecondaryHeader2ndStep = "Checkout: Overview",
  SuccessThx = "Thank you",
}

export enum CheckoutPageErrorMsg {
  FirstName = "Error: First Name is required",
  LastName = "Error: Last Name is required",
  PostalCode = "Error: Postal Code is required",
}